\name{NAPPA_example_input}
\alias{NAPPA_example_input}
\docType{data}
\title{
Nanostring nCounter data matrix
}
\description{
An example data matrix.  The 2 empty rows at the top of the file have been removed prior to loading and the housekeeping genes have been indicated in the left-most column. 
}
\usage{data(NAPPA_example_input)}
\format{
  A data frame with 81 observations on the following 15 variables.  The first three columns contain information about the gene and codeset.  The remaining 12 columns contain the raw code counts for each of the 12 input samples.
}

\keyword{datasets}
