\name{NS.Dilution}
\alias{NS.Dilution}
\docType{data}
\title{
Nanostring nCounter data matrix
}
\description{
An example data matrix. The housekeeping genes have been indicated in the left-most column.
}
\usage{data(NS.Dilution)}
\format{
  A data frame with 513 observations on the following 27 variables.  The first three columns contain information about the gene and codeset.  The remaining 24 columns contain the raw code counts for each of the 24 input samples.
}

\keyword{datasets}
