\name{NS.Lung}
\alias{NS.Lung}
\docType{data}
\title{
Nanostring nCounter data matrix
}
\description{
An example data matrix. The housekeeping genes have been indicated in the left-most column.
}
\usage{data(NS.Lung)}
\format{
  A data frame with 513 observations on the following 53 variables.  The first three columns contain information about the gene and codeset.  The remaining 50 columns contain the raw code counts for each of the 50 input samples.
}

\keyword{datasets}
