\name{nsolver.Dilution}
\alias{nsolver.Dilution}
\docType{data}
\title{
Nanostring nCounter data matrix
}
\description{
A data matrix processed using the nSolver method.
}
\usage{data(nsolver.Dilution)}
\format{
  A data frame with 492 observations on the following 24 variables.
}

\keyword{datasets}
