\name{NAPPA-package}
\alias{NAPPA-package}
\docType{package}
\title{
NAPPA: NAnostring Pre-Processing Algorithm.
}
\description{
Performs postive control, negative control and housekeeping normalisation of Nanostring mRNA data.  Enables multiple cartridges from the same codeset to be normalised as a single dataset.
}
\details{
\tabular{ll}{
Package: \tab NAPPA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-04\cr
License: \tab GPL-2\cr
}

}
\author{
Author: Mark Wappett

Maintainer: Mark Wappett <mark.wappett@astrazeneca.com>
}
\keyword{ package }
